refactor: 테스트용 클래스들 한곳에 모음
