--- android-di/src/test/java/io/hyemdooly/androiddi/component/FakeActivity.kt
@@ -1,10 +1,0 @@
-package io.hyemdooly.androiddi.component
-
-import io.hyemdooly.androiddi.base.HyemdoolyActivity
-import io.hyemdooly.androiddi.element.FakeFormatterInActivity
-import io.hyemdooly.di.annotation.Inject
-
-class FakeActivity : HyemdoolyActivity() {
-    @Inject
-    lateinit var formatter: FakeFormatterInActivity
-}
