--- android-di/src/test/java/io/hyemdooly/androiddi/element/FakeRepository.kt
@@ -1,3 +1,0 @@
-package io.hyemdooly.androiddi.element
-
-interface FakeRepository
