--- android-di/src/test/java/io/hyemdooly/androiddi/element/FakeRepositoryInApplication.kt
@@ -1,3 +1,0 @@
-package io.hyemdooly.androiddi.element
-
-class FakeRepositoryInApplication(db: FakeDatabase) : FakeRepository
