--- android-di/src/test/java/io/hyemdooly/androiddi/module/FakeApplicationModule.kt
@@ -1,18 +1,0 @@
-package io.hyemdooly.androiddi.module
-
-import android.content.Context
-import io.hyemdooly.androiddi.element.FakeDatabase
-import io.hyemdooly.androiddi.element.FakeRepositoryInApplication
-import io.hyemdooly.di.Module
-import io.hyemdooly.di.annotation.Singleton
-
-class FakeApplicationModule(private val applicationContext: Context) : Module() {
-    @Singleton
-    fun provideContext() = applicationContext
-
-    @Singleton
-    fun provideDatabase() = FakeDatabase()
-
-    @Singleton
-    fun provideRepository(db: FakeDatabase) = FakeRepositoryInApplication(db)
-}
