--- di/src/test/kotlin/io/hyemdooly/di/FakeChildModule.kt
@@ -1,8 +1,0 @@
-package io.hyemdooly.di
-
-import io.hyemdooly.di.annotation.Singleton
-
-class FakeChildModule(parentModule: Module) : Module(parentModule) {
-    @Singleton
-    fun provideFakeRepository(dao: FakeDao): FakeRepository = FakeRepositoryImpl(dao)
-}
