--- di/src/test/kotlin/io/hyemdooly/di/FakeDao.kt
@@ -1,3 +1,0 @@
-package io.hyemdooly.di
-
-interface FakeDao
