--- di/src/test/kotlin/io/hyemdooly/di/FakeFamily.kt
@@ -1,3 +1,0 @@
-package io.hyemdooly.di
-
-class FakeFamily()
