--- di/src/test/kotlin/io/hyemdooly/di/FakeParentModule.kt
@@ -1,10 +1,0 @@
-package io.hyemdooly.di
-
-import io.hyemdooly.di.annotation.Singleton
-
-class FakeParentModule : Module() {
-    @Singleton
-    fun provideFakeDao(): FakeDao = FakeDaoImpl()
-
-    fun provideFamily(): FakeFamily = FakeFamily()
-}
