--- di/src/test/kotlin/io/hyemdooly/di/FakePerson.kt
@@ -1,3 +1,0 @@
-package io.hyemdooly.di
-
-class FakePerson(val name: FakeName)
