--- di/src/test/kotlin/io/hyemdooly/di/FakeRepository.kt
@@ -1,3 +1,0 @@
-package io.hyemdooly.di
-
-interface FakeRepository
