--- di/src/test/kotlin/io/hyemdooly/di/FakeRepositoryImpl.kt
@@ -1,3 +1,0 @@
-package io.hyemdooly.di
-
-class FakeRepositoryImpl(val dao: FakeDao) : FakeRepository
