--- di/src/test/kotlin/io/hyemdooly/di/FakeViewModel.kt
@@ -1,8 +1,0 @@
-package io.hyemdooly.di
-
-import io.hyemdooly.di.annotation.Inject
-
-class FakeViewModel(val repository: FakeRepository) {
-    @Inject
-    lateinit var dao: FakeDao
-}
